vcsrepo: Add cron jobs to the state repository class
Moves in the daily cron scripts from the core repo, it makes more sense
to keep these with the rest of the repo management code even if that
gets moved from here at some point.

This should also essentially take care of #5 as once these run once the
data will be present.